==> notebooks/02_data_transformation.ipynb <==
# Data Transformation
#### I.Primary data Wrangling : 
1. Handling Missing values
2. Duplicates
3. correct data type: like dates 
4. Eleminate unecessary columns like UDI , Product ID
#### II.Feature Engineering 
1. label encoding 
2. catgorical encoding 
3. Converting Tempature from kalvin to Celcius
4. Feature scaling
5. Oversampling


import os 
import pandas as pd
import numpy as np


%pwd

data  = pd.read_csv("../artifacts/raw.csv")
data.head()

data.info()

data.info()

Missing values 

data.isnull().sum()

duplicated

data.duplicated().sum()

Eliminate unecesseary columns


data.drop(['UDI', 'Product ID'], axis=1, inplace=True)
data.head()

Categorical Encoding

from sklearn.preprocessing import OrdinalEncoder
encoder = OrdinalEncoder(categories=[['L', 'M', 'H']])

data['Type'] = encoder.fit_transform(data[['Type']])
data.head()

def type_of_failure(row_name):
    if data.loc[row_name, 'TWF'] == 1:
        data.loc[row_name, 'type_of_failure'] = 'TWF'
    elif data.loc[row_name, 'HDF'] == 1:
        data.loc[row_name, 'type_of_failure'] = 'HDF'
    elif data.loc[row_name, 'PWF'] == 1:
        data.loc[row_name, 'type_of_failure'] = 'PWF'
    elif data.loc[row_name, 'OSF'] == 1:
        data.loc[row_name, 'type_of_failure'] = 'OSF'
    elif data.loc[row_name, 'RNF'] == 1:
        data.loc[row_name, 'type_of_failure'] = 'RNF'

data.apply(lambda row: type_of_failure(row.name), axis=1)
data['type_of_failure'].replace(np.NaN, 'no failure', inplace=True)
data.drop(['TWF', 'HDF', 'PWF', 'OSF', 'RNF'], axis=1, inplace=True)
data.head()

from sklearn.preprocessing import LabelEncoder
encoder = LabelEncoder()

data['type_of_failure'] = encoder.fit_transform(data['type_of_failure'])
data.head()

convert Temperature from kalvin to celcius

data['Air temperature [c]'] = data['Air temperature [K]'] - 273.15
data['Process temperature [c]'] = data['Process temperature [K]'] - 273.15
data.drop(['Air temperature [K]', 'Process temperature [K]'], axis=1, inplace=True)
data.head()

data["Machine failure"].nunique

Scaling

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()

scale_cols = ['Rotational speed [rpm]', 'Torque [Nm]', 'Tool wear [min]', 'Air temperature [c]', 'Process temperature [c]']

data_scaled = scaler.fit_transform(data[scale_cols])
data_scaled = pd.DataFrame(data_scaled)
data_scaled.columns = scale_cols

data.drop(scale_cols, axis=1, inplace=True)

data_scaled = pd.concat([data, data_scaled], axis=1)
data_scaled.head()

Oversampling

from imblearn.over_sampling import SMOTE

smote = SMOTE(sampling_strategy='auto',random_state=42)

X = data_scaled.drop('type_of_failure', axis=1)
y = data_scaled['type_of_failure']

X_resampled, y_resampled = smote.fit_resample(X, y)

df_sampled = pd.concat([X_resampled, y_resampled], axis=1)
df_sampled.head()

df_sampled.value_counts('type_of_failure')

%pwd

from sklearn.model_selection import train_test_split
train,test= train_test_split(df_sampled,test_size=0.2,train_size=0.8)
train.to_csv("/local/home/hadjmefm/Predictive-Maintenance-System/artifacts/train.csv",index=False)
test.to_csv("/local/home/hadjmefm/Predictive-Maintenance-System/artifacts/test.csv",index=False)

df_sampled.to_csv("../artifacts/transformed_data.csv",index=False)
