==> notebooks/03_training.ipynb <==
import os 
import pandas as pd 
import numpy as np 
import mlflow
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score, f1_score, classification_report
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.model_selection import GridSearchCV

%pwd

train = pd.read_csv("../artifacts/train.csv")
train.head()

test = pd.read_csv("../artifacts/test.csv")
test.head()

## 1- Predict Machine Failure 

x_train = train.drop(["Machine failure","type_of_failure"],axis=1)
y_train = train["Machine failure"]
x_test = test.drop(["Machine failure","type_of_failure"],axis=1)
y_test = test["Machine failure"]

# Best parameters found from Grid Search
best_params = {
    'max_depth': 15,
    'min_samples_leaf': 1,
    'min_samples_split': 2,
    'n_estimators': 50
}

#Model with the best parameters
model = RandomForestClassifier(
    max_depth=best_params['max_depth'],
    min_samples_leaf=best_params['min_samples_leaf'],
    min_samples_split=best_params['min_samples_split'],
    n_estimators=best_params['n_estimators'],
    random_state=42
)

# Perform cross-validation
cv_scores = cross_val_score(model, x_train, y_train, cv=5)
print("Cross-validation scores:", cv_scores)
print("Mean cross-validation score:", cv_scores.mean())

# Train the model
model.fit(x_train, y_train)

# Evaluate the model
train_accuracy = model.score(x_train, y_train)
val_accuracy = accuracy_score(y_test, model.predict(x_test))

print("Training accuracy:", train_accuracy)
print("Validation accuracy:", val_accuracy)

### default RandomForest Model

rf = RandomForestClassifier()
mlflow.set_experiment("Faillure machine Model Evaluation ")
with mlflow.start_run():
  rf.fit(x_train, y_train)
  train_acc = rf.score(x_train, y_train)
  mlflow.log_metric("train_acc",train_acc)
  y_pred = rf.predict(x_test)
  eval_acc = accuracy_score(y_test, y_pred)
  mlflow.log_metric("eval_acc",eval_acc)
  prec = precision_score(y_test, y_pred)
  mlflow.log_metric("precision",prec)
  rec = recall_score(y_test, y_pred)
  mlflow.log_metric("recall",rec)
  f1 = f1_score(y_test, y_pred)
  mlflow.log_metric("f1_score",f1)
  mlflow.sklearn.log_model(rf,"default_randomforest_model")
mlflow.end_run()
print("training accuracy",train_acc) 
print("eval accuracy",eval_acc)

  

**The 100% training accuracy strongly indicates model overfitting.**

**This conclusion is further supported by the evaluation accuracy being marginally lower than the training accuracy**

### RandomForest with Hyperparmater Tuning 

data = pd.read_csv("../artifacts/transformed_data.csv")
X= data.drop(["Machine failure","type_of_failure"],axis=1)
y=data["Machine failure"]



param_grid = {
    'max_depth': [5, 10, 15],
    'min_samples_split': [2, 5, 10],
    'min_samples_leaf': [1, 2, 4],
    'n_estimators': [50, 100, 200]
}

grid_search = GridSearchCV(estimator=RandomForestClassifier(), param_grid=param_grid, cv=5, n_jobs=-1, verbose=2)
grid_search.fit(x_train, y_train)
mlflow.sklearn.log_model(grid_search.best_estimator_,"best randomforest model")

for k in grid_search.best_params_.keys():
    mlflow.log_param(k,grid_search.best_params_[k])
print("Best parameters found: ", grid_search.best_params_)
print("Model saved in run %s" % mlflow.active_run().info.run_uuid)

### Model with best params


#split
X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.2, random_state=42)

# Best parameters found from Grid Search
best_params = {
    'max_depth': 15,
    'min_samples_leaf': 1,
    'min_samples_split': 2,
    'n_estimators': 50
}

#Model with the best parameters
model = RandomForestClassifier(
    max_depth=best_params['max_depth'],
    min_samples_leaf=best_params['min_samples_leaf'],
    min_samples_split=best_params['min_samples_split'],
    n_estimators=best_params['n_estimators'],
    random_state=42
)

# Perform cross-validation
cv_scores = cross_val_score(model, X_train, y_train, cv=5)
print("Cross-validation scores:", cv_scores)
print("Mean cross-validation score:", cv_scores.mean())

# Train the model
model.fit(X_train, y_train)

# Evaluate the model
train_accuracy = model.score(X_train, y_train)
val_accuracy = accuracy_score(y_val, model.predict(X_val))

print("Training accuracy:", train_accuracy)
print("Validation accuracy:", val_accuracy)

1. **The cross-validation scores are consistently high, indicating stable model performance across different subsets of the data.**
2. **The mean cross-validation score is very close to the individual fold scores, reinforcing the model's reliability.**
3. **The training accuracy is very high, suggesting effective learning from the training data.**
4. **The validation accuracy is also high, indicating good generalization to new, unseen data.**

from sklearn.metrics import classification_report, roc_auc_score

# Predict probabilities for ROC-AUC
y_pred_proba = model.predict_proba(X_val)[:, 1]

# Classification report
print(classification_report(y_val, model.predict(X_val)))

# ROC-AUC score
roc_auc = roc_auc_score(y_val, y_pred_proba)
print("ROC-AUC Score:", roc_auc)
